third_party/WebKit/LayoutTests/http/tests/devtools/console/console-error-on-call-frame-expected.txt

CONSOLE ERROR: line 1: 42
Tests that console.error does not throw exception when executed in console on call frame.

Set timer for test function.
Script execution paused.
Script execution resumed.
console.error(42)
VM:1 42
(anonymous) @ VM:1
testFunction @ console-error-on-call-frame.html:10
setTimeout (async)
scheduleTestFunction @ debugger-test.js:3
(anonymous) @ VM:1
undefined
